// comonl-comonlwcl/src/app/modules/comonlapi/model/formRicercaComunicazione.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Provincia } from './provincia';
import { StatoComunicazione } from './statoComunicazione';
import { TipoComunicazione } from './tipoComunicazione';
import { TipoComunicazioneTu } from './tipoComunicazioneTu';
import { TipoTrasferimento } from './tipoTrasferimento';


export interface FormRicercaComunicazione { 
    codiceFiscaleAzienda?: string;
    denominazioneAzienda?: string;
    codiceFiscaleLavoratore?: string;
    cognomeLavoratore?: string;
    nomeLavoratore?: string;
    dataInvioInserimentoDa?: Date;
    dataInvioInserimentoA?: Date;
    statoComunicaziones?: Array<StatoComunicazione>;
    codiceRegionale?: string;
    tipoComunicazione?: TipoComunicazione;
    tipoComunicazioneTu?: TipoComunicazioneTu;
    provinciaProtocollo?: Provincia;
    annoProtocollo?: number;
    numeroProtocollo?: number;
    codiceFiscaleSoggetto?: string;
    tipoTrasferimento?: TipoTrasferimento;
    tipoTracciato?: string;
    tipoVariazione?: string;
}